api/requirements.txt
Flask
flask-cors
openai-whisper
